Added Shiny app files and documentation  - App.R: Main application file. - app_server.R: Server-side logic of the app. - app_ui.R: User interface components of the app. - bank-full-2025.csv: Dataset used in the application. - ReadMe.pdf: Documentation and project overview.  This update includes the complete Shiny app code and a detailed README file.
This commit contains the essential files for the predictive analysis of marketing campaign success in the banking sector. 
The README provides an overview, features, and steps to run the app locally. The Shiny app utilizes LDA for prediction and 
includes interactive data visualization using Plotly and DT tables.

The deployed Shiny app is available online at:
https://mohammed-saif-alotaibi.shinyapps.io/fainlproject/